Refactor escape column methods
