+++ src/main/scala/com/amazon/deequ/suggestions/ConstraintSuggestionRunner.scala
@@ -16,7 +16,6 @@
 
 package com.amazon.deequ.suggestions
 
-import com.amazon.deequ.analyzers.Preconditions.escapeColumn
 import com.amazon.deequ.analyzers.{DataTypeInstances, KLLParameters}
 import com.amazon.deequ.{VerificationResult, VerificationSuite}
 import com.amazon.deequ.checks.{Check, CheckLevel}
@@ -24,6 +23,7 @@ import com.amazon.deequ.io.DfsUtils
 import com.amazon.deequ.profiles.{ColumnProfile, ColumnProfilerRunner, ColumnProfiles}
 import com.amazon.deequ.repository.{MetricsRepository, ResultKey}
 import com.amazon.deequ.suggestions.rules._
+import com.amazon.deequ.utilities.ColumnUtil.escapeColumn
 import org.apache.spark.annotation.Experimental
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.{DataFrame, SparkSession}
